requirements.txt
psycopg2~=2.9.1
python-decouple~=3.4